+++ packages/react/src/composite/SegmentedControl/src/contexts/SegmentedControlContext/index.ts
@@ -1,0 +1,5 @@
+export {
+  SegmentedControlContext,
+  useSegmentedControlContext,
+} from "./SegmentedControlContext";
+export type { SegmentedControlContextValue } from "./SegmentedControlContext.types";